Rename `skillsData.js` -> `rawSkillsData.js`
