+++ src/components/skills/rawSkillsData.ts
